fix: made it so the client only requests for mfa options on load
